+++ TMS.API/TMS.API.csproj
@@ -240,7 +240,6 @@
     <Compile Include="Areas\HelpPage\SampleGeneration\SampleDirection.cs" />
     <Compile Include="Areas\HelpPage\SampleGeneration\TextSample.cs" />
     <Compile Include="Areas\HelpPage\XmlDocumentationProvider.cs" />
-    <Compile Include="Classes\AzureStorageMultipartFormDataStreamProvider.cs" />
     <Compile Include="Classes\CustomAuthorize.cs" />
     <Compile Include="Controllers\DriverController.cs" />
     <Compile Include="Controllers\MediaController.cs" />
