+++ TMS.BusinessGateway/Classes/AzureStorageMultipartFormDataStreamProvider.cs
@@ -5,9 +5,10 @@ using System.IO;
 using System.Linq;
 using System.Net.Http;
 using System.Net.Http.Headers;
-using System.Web;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace TMS.API.Classes
+namespace TMS.BusinessGateway.Classes
 {
     public class AzureStorageMultipartFormDataStreamProvider : MultipartFormDataStreamProvider
     {
