+++ TMS.BusinessGateway/Gateway/TaskGateway.cs
@@ -10,7 +10,7 @@ using TMS.DomainGateway.Task.Interfaces;
 
 namespace TMS.BusinessGateway.Gateway
 {
-    public class TaskGateway: ITaskGateway
+    public class TaskGateway : ITaskGateway
     {
         public IDriverTask DriverTask
         {
@@ -68,7 +68,12 @@ namespace TMS.BusinessGateway.Gateway
         }
         public IInvoiceTask InvoiceTask
         {
-            get { return new  BusinessInvoiceTask(new Invoice()); }
+            get { return new BusinessInvoiceTask(new Invoice()); }
+        }
+
+        public IMediaTask MediaTask
+        {
+            get { return new BusinessMediaTask(); }
         }
     }
 }
