+++ TMS.DomainGateway/Gateway/TaskGateway.cs
@@ -23,5 +23,6 @@ namespace TMS.DomainGateway.Gateway
         public abstract ITripTask TripTask { get; set; }
         public abstract IReportTask ReportTask { get; set; }
         public abstract IInvoiceTask InvoiceTask { get; set; }
+        public abstract IMediaTask MediaTask { get; set; }
     }
 }
