+++ TMS.DomainGateway/TMS.DomainGateway.csproj
@@ -50,6 +50,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Task\Authenticate.cs" />
     <Compile Include="Task\Interfaces\IInvoiceTask.cs" />
+    <Compile Include="Task\Interfaces\IMediaTask.cs" />
     <Compile Include="Task\Interfaces\IReportTask.cs" />
     <Compile Include="Task\Interfaces\ITripTask.cs" />
     <Compile Include="Task\Interfaces\IPoolTask.cs" />
@@ -59,6 +60,7 @@
     <Compile Include="Task\Interfaces\IVehicleTask.cs" />
     <Compile Include="Task\Master.cs" />
     <Compile Include="Task\Invoice.cs" />
+    <Compile Include="Task\Media.cs" />
     <Compile Include="Task\Report.cs" />
     <Compile Include="Task\Trip.cs" />
     <Compile Include="Task\Pool.cs" />
